// src/interfaces/sale-order/saleOrder.ts
import { ICurrency } from "../currency/currency";
import { ICustomer } from "../customer/customer";
import { IEstablishment } from "../establishment/establishment";
import { IPointOfSale } from "../establishment/pointOfSale";
import { IUser } from "../user/user";
import { IWarehouse } from "../warehouse/warehouse";
import { ISaleOrderProduct } from "./product";
import { ISaleOrderStatus } from "./status";

export interface ISaleOrder {
    id: number;
    customer_id: number;
    currency_id: number;
    point_of_sale_id: number;
    warehouse_id: number;
    seller_id: number;
    status_id: number;
    amount: number;
    ordered_at: string;
    expires_at: string | null;
    billed_at: string | null;
    name: string;
    identification_document: string;
    phone: string | null;
    email: string | null;
    address: string | null;
    comments: string | null;
    created_by: number | null;
    updated_by: number | null;
    created_at: string;
    updated_at: string;
    customer: ICustomer;
    currency: ICurrency;
    establishment: IEstablishment;
    point_of_sale: IPointOfSale;
    warehouse: IWarehouse;
    seller: IUser;
    status: ISaleOrderStatus;
    products: ISaleOrderProduct[];
};